Rename deploy package file name and fix comments
