+++ internal/pkg/deploy/pipeline_test.go
@@ -1,7 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-// Package deploy holds the structures to deploy infrastructure resources.
 package deploy
 
 import (
